Made the code compiled on VS2008 (but the memory debugger cannot be used right now since it used an unordered_map) The code has been also tested on AMD
git-svn-id: http://yats.googlecode.com/svn/trunk@44 d59e3cf1-15ef-577c-704d-36b40c253874

+++ src/sys/tasking.cpp
@@ -389,7 +389,7 @@ namespace pf {
               << double(chunkNum * TaskStorage::chunkSize) / 1024
               << "KB" << std::endl;
 #endif /* PF_TASK_STATICTICS */
-    uint32 allocateNum = 0;
+    int64 allocateNum = 0;
     for (size_t i = 0; i < threadNum; ++i)
       allocateNum += this->local[i].allocateNum;
     FATAL_IF (allocateNum < 0, "** You may have deleted a task twice **");
